Extract XML creation to separate file
